feat(UI-1107): invite user to organization - refactor
